cyber_security_ai_project_pro_grade/requirements.txt
fastapi
uvicorn[standard]
confluent-kafka
sqlalchemy
psycopg2-binary
pydantic
scikit-learn
joblib
structlog
prometheus-client
starlette-exporter
python-jose[cryptography]
requests
alembic
pytest
pytest-asyncio
mlflow
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-requests
shap
